Make TagsEditor to accept only `onRemoveTags`
Instead of keeping two versions of pretty much the same code for removing
tags, we've picked one that can handle all the cases by considering the removement to always operate on multiple entities. In addition, we fixed
a bug where tags could be removed even if the input is not empty, which
goes against the requirements and common sense.